+++ app/services/positions/importing/tinkoff_service.rb
@@ -28,7 +28,8 @@ module Positions
         @portfolio = portfolio
 
         set_initial_values
-        parse_xls_file
+        set_header_indeces
+        parse_xls_file(data_sheet_index: 0)
         save_positions
       end