+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - sitemap
+- Freedom Finance broker importing data
 
 ### Modified
 - use cache for storing quote coupons sum for time range